Refactor file arquitecture
