--- movie/source.rb
@@ -1,15 +1,0 @@
-class Source
-  attr_accessor :id, :name
-  attr_reader :items
-
-  def initialize(id, name)
-    @id = id
-    @name = name
-    @items = []
-  end
-
-  def add_item(item)
-    @items << item
-    item.source = self
-  end
-end
