+++ src/music_album/preserve_music_albums.rb
@@ -2,10 +2,10 @@ require 'json'
 
 class PreserveMusicAlbums
   def gets_music_albums
-    return unless File.exist?('./music_albums.json')
+    return unless File.exist?('./src/music_album/music_albums.json')
 
     saved_music_albums = []
-    file = File.read('./music_albums.json')
+    file = File.read('./src/music_album/music_albums.json')
     data_hashes = JSON.parse(file)
     data_hashes.each do |music_album|
       saved_music_albums << MusicAlbum.from_hash(music_album)
@@ -17,6 +17,6 @@ class PreserveMusicAlbums
     return if music_albums.empty?
 
     data_hashes = music_albums.map(&:to_hash)
-    File.write('./music_albums.json', JSON.pretty_generate(data_hashes))
+    File.write('./src/music_album/music_albums.json', JSON.pretty_generate(data_hashes))
   end
 end
